Refactor environment validation and refine Bluetooth reconnection logic
- Split `lib/env.ts` into `clientEnvSchema` and `serverEnvSchema` to enforce strict validation of public variables on the client while gracefully handling missing server secrets.
- Consolidate Bluetooth constants and messages into `constants/bluetooth-reconnection.ts`, removing `constants/bluetooth-messages.ts`.
- Update `hooks/useBluetoothHRM.ts` to use `env` from `lib/env.ts`, removing direct `process.env` access and verbose comments.
- Implement counter-based "Connection Storm" detection (3 disconnects in 10s) instead of a simple threshold.
- Ensure explicit `disconnect()` call in `useBluetoothHRM` cleanup to prevent background connection leaks.
- Update unit tests for environment validation and Bluetooth constants.
