+++ app/client/connect/ConnectView.tsx
@@ -19,7 +19,7 @@ import WorkoutControls from './WorkoutControls'
 import ResetSection from './components/ResetSection'
 import { useState, useEffect } from 'react'
 import logger from '@/utils/logger'
-import { BLUETOOTH_MESSAGES } from '@/constants/bluetooth-messages'
+import { BLUETOOTH_MESSAGES } from '@/constants/bluetooth-reconnection'
 import { WorkoutStatus } from '../../../types/workout'
 import { UserProfileState, HrZoneData } from '@/types/connect'
 import { BluetoothConnectionStatus } from '@/types/bluetooth'
